backend/requirements.txt
fastapi
python-dotenv
google-genai
openpyxl
python-multipart
uvicorn